backend/pyproject.toml
[project]
name = "nora-chatbot"
version = "0.1.0"
description = "Intelligent Atlan Support Chatbot using LangGraph and FastAPI"
readme = "README.md"
authors = [
    { name = "jannatkhandev", email = "jannatkhandev@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "langgraph>=0.6.7",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "langchain-community>=0.3.29",
    "pinecone>=7.3.0",
    "sentence-transformers>=5.1.0",
    "pydantic>=2.11.8",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "bs4>=0.0.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
    "pytest-asyncio>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
